+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "alloglot" extension will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adhere's to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [2.0.1] - 2024-01-30
 
 - Move to a new repository.
 - use a different name for each language's diagnostics collection